refactor(starknet_api): move `GlobalContractCache` from `blockifier`
To be used by `ClassManager`, which should not depend on `blockifier`.
Also, it's standalone.
Code move only, no changes.
